TP5/requirements.txt
gymnasium[box2d]>=1.0.0
stable-baselines3>=2.0.0
Pillow>=9.0.0
